test: Move shared test data classes from main to test.
- Clean up SharedTestDataADM.
- Update docs.
