--- webapi/src/main/scala/org/knora/webapi/sharedtestdata/BUILD.bazel
@@ -1,15 +1,0 @@
-package(default_visibility = ["//visibility:public"])
-
-load("@io_bazel_rules_scala//scala:scala.bzl", "scala_library")
-
-scala_library(
-    name = "sharedtestdata",
-    srcs = glob(["*.scala"]),
-    unused_dependency_checker_mode = "warn",
-    deps = [
-        "//webapi/src/main/scala/org/knora/webapi",
-        "//webapi/src/main/scala/org/knora/webapi/messages",
-        "@maven//:io_spray_spray_json_2_12",
-        "@maven//:com_typesafe_akka_akka_http_spray_json_2_12",
-    ],
-)
